+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - de-duplicate features when creating a features gdf (#1400)
 - validate tolerance when consolidating intersections (#1401)
+- fix simplify_graph treating equally-missing edge_attrs_differ values as differing (#TBD)
 
 ## 2.1.1 (2026-07-21)